12.reinforcement_learning_with_mario_bros/2023/hard_coding/requirements.txt
gym-super-mario-bros
sewar

numpy
opencv-python
Pillow

tensorflow
h5py
tensorflow_hub
# torch
# torchvision
# torchaudio